requirements.txt
Django==1.11.6
pytz==2017.2
RPi.GPIO==0.6.3
